+++ src/main/java/com/codessquad/qna/HomeController.java
@@ -12,21 +12,41 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
+import java.util.ArrayList;
+import java.util.List;
 @Controller
 public class HomeController {
     private Logger logger = LoggerFactory.getLogger(HomeController.class);
 
     @Autowired
     private QuestionRepository questionRepository;
+    private List<Page> pages;
+    private static final int SIZE_OF_QUESTIONS_OF_PAGE = 15;
 
     @GetMapping("/")
     public String index(Model model) {
-        PageRequest pageRequest = PageRequest.of(1, 5, Sort.by("createdDate").descending());
-        Page<Question> pages = questionRepository.findAll(pageRequest);
-        model.addAttribute("questions", questionRepository.findAll(pageRequest));
-        model.addAttribute("countOfPages", pages.getContent());
-        logger.info("pages.getContent : {}" , pages.getContent().size());
+        pages = createPages();
+        model.addAttribute("questions", questionRepository.findAll());
+        model.addAttribute("pages", pages);
         return "index";
     }
 
+    @GetMapping("/test1")
+    public Page viewQuestionList() {
+        pages = createPages();
+        return pages.get(0);
+    }
+
+    private List<Page> createPages() {
+        int numberOfQuestions = questionRepository.findAll().size();
+        int lastPage = numberOfQuestions / SIZE_OF_QUESTIONS_OF_PAGE;
+        List<Page> pages = new ArrayList<>();
+        PageRequest pageRequest;
+
+        for (int i = 0; i < lastPage; i++) {
+            pageRequest = PageRequest.of(i, 5);
+            pages.add(questionRepository.findAll(pageRequest));
+        }
+        return pages;
+    }
 }
