+++ src/main/resources/import.sql
@@ -8,4 +8,16 @@ INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE,
 INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (7, 'CONTENTS2', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(),FALSE ,'6번글', 2, 0);
 INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (8, 'CONTENTS2', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(),FALSE ,'7번글', 2, 0);
 INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (9, 'CONTENTS2', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(),FALSE ,'8번글', 2, 0);
+INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (10, 'CONTENTS2', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(),FALSE ,'8번글', 2, 0);
+INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (11, 'CONTENTS2', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(),FALSE ,'8번글', 2, 0);
+INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (12, 'CONTENTS2', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(),FALSE ,'8번글', 2, 0);
+INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (13, 'CONTENTS2', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(),FALSE ,'8번글', 2, 0);
+INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (14, 'CONTENTS2', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(),FALSE ,'8번글', 2, 0);
+INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (15, 'CONTENTS2', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(),FALSE ,'8번글', 2, 0);
+INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (16, 'CONTENTS2', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(),FALSE ,'8번글', 2, 0);
+INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (9, 'CONTENTS2', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(),FALSE ,'8번글', 2, 0);
+INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (9, 'CONTENTS2', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(),FALSE ,'8번글', 2, 0);
+INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (9, 'CONTENTS2', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(),FALSE ,'8번글', 2, 0);
+INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (9, 'CONTENTS2', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(),FALSE ,'8번글', 2, 0);
+INSERT INTO QUESTION (ID, CONTENTS, CREATED_DATE, MODIFIED_DATE, DELETED, TITLE, WRITER_ID, COUNT_OF_ANSWER) VALUES (9, 'CONTENTS2', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(),FALSE ,'8번글', 2, 0);
