+++ src/main/resources/templates/index.html
@@ -34,14 +34,14 @@
                   <ul class="pagination center-block" style="display:inline-block;">
 
                       <li><a href="#">«</a></li>
-                      {{#countOfPages}}
+                      {{#pages}}
                       <li><a href="#">1</a></li>
-                      {{/countOfPages}}
+                      {{/pages}}
 <!--                      <li><a href="#">2</a></li>-->
 <!--                      <li><a href="#">3</a></li>-->
 <!--                      <li><a href="#">4</a></li>-->
 <!--                      <li><a href="#">5</a></li>-->
-<!--                      <li><a href="#">»</a></li>-->
+                      <li><a href="#">»</a></li>
                 </ul>
               </div>
               <div class="col-md-3 qna-write">
